Added hostname range expansion functionality for -w option. Added -X and -x options to control this feature.
Hostname ranges can be of various forms like host1-5 host01-05 or host1-host5.
The range expansion function tries to do its best to detect whether a range
is meant in cases of internal range operators ('-' by default), but this is
easily foiled. Users can set PDSH_RANGE_OPERATOR to change the default operator
in the case of hostnames such as rtr1-2 (which would, by default, be expanded
to rtr1,rtr2) Also, setting PDSH_RANGE_OPERATOR to "", or passing the -x
option to pdsh will disable range expansion. The range operator may be set
on the command line via the -X option (e.g. -X:).

Currently, the expansion function will assume that a single hostname was
meant if it encounters any error while attempting to expand the range.