From alan keane at Quadrics: I've noticed that the configure script dosn't seem to find the pthread library correctly if running on an OSF / TRU64 system.
I've included a patch to configure.ac which I believe fixes the problem,
just a one liner but it does allow a TRU64 ./configure to find the pthread
library. I've tested the new configure on Linux and TRU64 but I havn't tried
it on AIX or other systems.

The patch allows pdsh to build on a TRU64 box as long as GCC is the default
compiler.
The OSF cc compiler is a bit too strict and ends up failing a lot of the
AC_* macros!