pdcp doesn't parallelize rcp operation
Don't use list_for_each to traverse list of files in pdcp threads,
as the this list is shared between threads, and list_for_each
takes the list lock for the duration of the traversal (including
while functions are in progress). The use of list_for_each
effectively serialized all pdcp invocations.