Handle errors from XSI-compliant strerror_r() in xstrerrorcat()
On some recent Linux platforms strerror_r() will return -1
with errno set to EINVAL for a couple of values of errno. This
resulted in an uninitialized error string being used in
xstrerrorcat(). This fix handles EINVAL for this version
of strerror_r() and appends the string "Unknown error NN"
to match the more typical behavior.