torque: Fix segfault with no default PBS server
Apparently, pbs_server may be NULL after the call the pbs_connect()
if no default pbs server is configured and PBS_DEFAULT is not set.
This resulted in a segv when testing the torque module on a system
with libtorque installed but no torque configuration.

This patch fixes the error message to avoid using pbs_server if it
is NULL, and to give a hint as to (possibly) why pbs_connect() failed
e.g.:

./pdsh -j 0 hostname
pdsh@wednesday: Failed to connect to torque server: PBS_DEFAULT not set or no default server