Fix some incorrect uses of inactive modules
There are some places where inactive / uninitialized modules
are being incorrectly used and/or referenced in the pdsh mod.c code.
This patch introduces a wrapper for list_next() on a module list
iterator that skips inactive/uninitialized modules, and replaces
list_next() with this function in a couple key places.

Also, improve the mod_get_module_names() implementation to be
simpler. Update one comment for clarity.