Switch to rresvport_af for protocol independence
Switch to rresvport_af() in privsep_resvport implementation in order
to support AF_INET6 in the future. Introduce privsep_rresvport_af()
which takes a protocol family argument, and implement privsep_rresvport()
as privsep_rresvport(AF_INET) for compatibility with older code.

The family (AF_INET or AF_INET6) is encoded in the top 16 bits of
the lport sent to the privsep server. This should be fine since the
highest expected reserved port is < 1024.