Preserve argument list in ssh rcmd module
Pass remote command argument list to ssh in the same form
it was passed to pdsh, instead of collapsing all args into
a single command string.

For example, when pdsh was called like so:

  pdsh -Rssh -w host0 cmd arg1 arg2 ...

it would invoke ssh like this:

  ssh -2 -a -x -l user host0 "cmd arg1 arg2 .."

i.e. ssh will see cmd and its args as a single argument. After
this fix, ssh would be invoked as:

 ssh -2 -a -x -l user host0 cmd arg1 arg2 ..

i.e ssh will see cmd and it args as separate arguments. (Since
pdsh exec's ssh, these args are not subject to shell expansion
again until they are invoked by the shell on the remote host)

This change should only make a difference in cases where there
is a difference between

 ssh host "cmd args..."

and

 ssh host "cmd" "arg1" "arg2" ...