Generate an error for missing genders file only when genders option is used
Do not generate an error for a missing genders file unless one of
the genders options is explicitly used, i.e. -F -a -g -X or -i.
In the case one of these options was used and the genders data
can not be read, then exit immediately with an error.

This is probably clearer than the current behavior of generating
a non-fatal error every time, even when genders module isn't doing
much of anything.