testsuite: Add pcptest rcmd module for use in testing {r}pdcp
To assist writing standalone pdcp tests, the pcptest rcmd module
wraps the invocation of the pcp server in a chdir which simulates
activity on remote hosts. Tests may then compare results in one
or several subdirs to verify the efficacy of the test. The pcptest
module invokes the equivalent of

  sh -c "cd %h; <pdsh pdcp server command>"

where %h is expanded to the target host, as in the exec module,
and <pdsh pdcp server command> is the pdcp -z command normally
generated by pdsh to invoke the pdcp server through the configured
remote command protocol (ssh, rsh, etc)