testsuite: Add SANITY prereq to tests in non-writable dir
The final test of t1000-dshbak.sh tests for graceful failure
when using -d with a non-writeable directory. Add the SANITY
prereq to this tests to ensure we skip it when file perms
aren't be adhered to in a sane fashion.

Update issue 11
Possible fix for failure in t1000-dshbak.sh