testsuite: Add tests for rcmd/ssh module
Add simple tests for the rcmd/ssh pdsh module. These tests
do not require that ssh be actually installed or functional, but
instead make use of a script that prints the arguments that would
have been given to the ssh executable.

Update Issue 12
Added test to pdsh testuite for {r}pdcp functionality with rcmd/ssh.