Fix pdcp when using ssh as connect method
The change introduced in r1262 broke the use of r/pdcp with
rcmd/ssh because the ssh module used pdsh_remote_argv() instead
of `cmd' to create the remote command line.

In the case of r/pdcp, we need to use `cmd' passed into sshcmd()
because this commandline has been updated to run the pdcp server
instead of any remote command line specified by the user (as in
normal pdsh).

Resolves Issue 12.