Ensure linker is passed -E flag on HP-UX
User reports in google code issue 18 that symbols are not being
properly exported from the pdsh executable (specifically rcmd_opt_set),
which is causing rcmd modules to fail to load.

The user found that the  addition of -Wl,-E to LDFLAGS resolved the
problem, so we add a special case for that in configure.

Resolves Issue 18