genders: change genders_query handling to compile-time
The hack to check for genders_query(3) support in libgenders
at runtime using dlopen/dlsym was causing problems on some systems.
This is because the two reasonable methods for getting a handle
to the global or libgenders symbols weren't working in all cases.
That is

  dlopen(NULL)

to get a global symbol table handle wasn't supported on all systems
and

  dlopen("libgenders.so")

doesn't work for systems that only install libgenders.so as part
of a genders development package (otherwise they only install the
version-numbered shared libraries and not the .so link)

So, remove the use of dlopen/dlsym in the genders module in favor
of compile time use of HAVE_GENDERS_QUERY.

Resolves Issue 1.

Update Issue 21

This commit should fix the second of the two issues open against
the genders module on HP-UX.