fix compilation of pdsh with static modules
The use of --enable-static-modules was not properly tested after
the module load code was redone to allow conflicting modules to
be loaded simultaneously by pdsh. The main problem appears to be
that some forward declarations were ifdef'd out during a static
module build. This commit fixes that issue.

Update Issue 23

Move prototypes in mod.c so they are not ifdef'd out when building
with --enable-static-modules. This should fix the static module
build.