testsuite: add DYNAMIC_MODULES prereq
Some tests depend on dynamic module support being built in to pdsh,
i.e. they will fail with --enable-static-modules. Check for
static-modules build in the testsuite and set the DYNAMIC_MODULES
prereq if a static-modules build is not detected. Require
DYNAMIC_MODULES for tests that need it.

Update Issue 23

Fixes for testsuite and --enable-static-modules