more dynamic max username length
The ruser and luser username variables use statically defined buffers
with an  arbitrary, hard-coded size of 17. While this length was
is sufficient for most cases, user Scott M. McDermott rightly suggested
that this arbitrary limit was incorrect and was causing issues at
his site.

The most portable way to handle username length appears to be the
use of sysconf (_SC_LOGIN_NAME_MAX), and since this is determined
at runtime, we need to make opt->{l,r}user dynamically allocated at
program startup.

Also, it makes no sense to continue if the username has been truncated,
so if we see that a username is going to overflow _SC_LOGIN_NAME_MAX,
abort the program with a useful error.

Resolves Issue 24