testsuite: tests for optional %u and %h in PDSH_SSH_ARGS
Before the refactor of sshcmd to use pipecmd(), PDSH_SSH_ARGS was
not required to contain %u and %h, instead the args always ended
with "-l%u %h". Add tests to ensure this older usage continues to
work.

Update Issue 25

Tests for optional use of %u and %h in PDSH_SSH_ARGS.