ssh: do not require %u and %h in PDSH_SSH_ARGS
As described in issue 25, the use of PDSH_SSH_ARGS changed
after the refactor of sshcmd.c to use pipecmd(). Previously,
the user was not required to employ the %u and %h parameters
to insert the username and hostname into the ssh cmdline.

In order to preserve this older behavior, and also allow
PDSH_SSH_ARGS to override the entire ssh commandline, scan
the argument list for %h and %u and append them only if needed.

Resolves Issue 25.