dshgroup: allow dshgrou path override at runtime
It will be useful for testing and perhaps in other cases to override
the default path the dshgroup module searches for group files at
runtime, instead of only from ./configure. This patch adds support for
a DSHGROUP_PATH environment variable that can be set to a directory
in which to search for group files, if none exist in the user's homedir.