Enhance the interface to read wcoll files
Add a new interface to read wcoll files from an optional
colon separated path:

 hostlist_t read_wcoll_path (const char *path, const char *file);

If 'file' is not an absolute path or relative to '.' or '..', then
read_wcoll_path will search 'path' for 'file'.

This new interface also supports including wcoll files from other
files via a '#include' directive, e.g.

 #include 'newfile'

(The # must be the first character of the line in question). The
same path will be searched to resolve any included files.