testsuite: add tests for pdsh truncation of output
Update Issue 36

Add a test that reproduces the user observation that pdsh is truncating
very long lines. Lines of length > 8191 characters are probably required
to reproduce this behavior, which appears to be due to a static buffer
used in pdsh output routines, along with a maximum circular buffer size
of 8192 characters. Both issues will need to be resolved before this
test passes.