Increase output cbuf max size to 128K
The maximum size of output cbufs (circular buffers) was
limited to 8K at creation time. This effectively limited the
maximum line length to 8191 characters when there were multiple
lines of output, because the cbuf by default is set to wrap
and thus discard data. Since a maximum size is required, set it
to 128K for now, with the option to handle overwriting better
in the future.