Use a dynamic buffer for output to avoid truncation
Dynamically determine the length of lines to be printed during output
and allocate a buffer big enough for entire line. This avoids possible
truncation due to a statically sized buffer. The fix uses cbuf_peek_line()
to determine next line size (if any), then creates a buffer for the output
based on the line size.

A better fix in the future would be to avoid using an intermediary buffer
at all and directly send output from the cbuf to the output file descriptors.

Resolves Issue 36