Do not print empty lines in pdsh output
In commit c37f0b533b93c0d6825df83dc3035ec54a1ce3e8
"Use a dynamic buffer for output to avoid truncation" a test
for empty lines was dropped, thus resulting in empty lines
printed from pdsh in some cases. This was easily recreated
by running rsh/mrsh with -S and a command with no output.
The return code was stripped from the empty line of output,
but since there was still an output buffer, the empty line
was printed with confusing output.

This commit replaces the test for line length > 0.

Resolves Issue 54.