genders: Fix redundant traversal of wcoll with -g
In commit b9fbf059627d278c1391e316da1525b82b700f0e code was added
to filter hostlists provided by -w via a genders query given in
the -g option. Unfortunately, when using -g alone, the same code
was (redundantly) used. This resulted in an unecessary walk of
the entire hostlist, host by host, to apply the genders query
which originally generated the hostlist in the first place!

This commit fixes the redundant hostlist walk by having the genders
module "consume" attributes as they are used to generate the
wcoll in the first place. Thus, if there are any attributes/query
strings remaining at "post option processing", then we know these
were not used to generate the original wcoll, and we need to
apply them to filter the existing wcoll hostlist.

Resolves Issue 59.