genders: Speed up genders_filter
genders_filter was filtering hosts from the wcoll by expanding
the hostlist and calling genders_testquery on each individual host.
This was extremely slow for many hosts. Instead, this patch gets
the list of hosts matching each genders query in turn, and gets
the intersection of each list with the wcoll, then takes the union
of each such intersection to imply an overall OR.