pdsh timeouts can be set in environment variables
pdsh allows environment variables for various settings. However, the connect
and command timeouts only have command line options.

This change makes it so these timeouts can be specified as environment
variables. That way, if a site decides it is best, the variables can be
set to impact pdsh commands that are buried in infrastructure.

PDSH_CONNECT_TIMEOUT
and
PDSH_COMMAND_TIMEOUT

Are the environment variables.