pdsh: use straight dlopen instead of libltdl
Dump libltdl in favor of straight dlopen/dlsym for pdsh modules.
The new libltdl wrapper is a whole thing of its own and a big burden
to use (it wants to be installed as a subproject inside of the
project which uses it)

This likely means pdsh dynamic modules are no longer supported on
systems. Likely those systems don't work now anyway. In the future
if an unlikely port is needed to a system requiring libltdl, it should
be possible to tackle re-implementation of libltdl (or some other
portability layer) at that time.