pipecmd: improve error for nonzero exit status
In pipecmd and error is emitted when waitpid(2) returns non-zero status
for a child process. However, the error only includes the result of
WEXITSTATUS(), which may be zero if the process was terminated due to
a signal. To avoid confusing error messages "Exited with exit code 0",
return a different error message if the process was killed.