dshbak: fix uninitialized use of $tag on empty input
Problem: The fix for hostnames of `0` in dshbak, i.e. the
following commit

 commit dee6adf334376da560afa0f8beebe4cae36d8e7c
 Author: Mark A. Grondona <mgrondona@llnl.gov>
 Date:   Thu Jun 25 13:37:51 2015 -0700

    dshbak: Handle hostname of "0"

    @garlick found that dshbak can't handle the smallest of numbers.
    This is because dshbak skips lines where $tag resolves to false.
    But in perl, 0 is also false.

    What we really want to check for is an "empty" tag, so fix up the
    test to reflect that.

    Resolves #70.

breaks handling of empty input to dshbak, since now the variable
"$tag" can be used when it is uninitialized. The proper fix for
issue #70 was to check if the variable was either uninitialized or
an empty string.

Fixes #132