cbuf: avoid unused variable errors in production build
Problem: Variables in cbuf.c are used solely for the purpose
of later use in assert(). When built with NDEBUG, this results
in a "set but unused" variable error from GCC.

Fix the error by addition of an _unused(x) macro which tricks
the compiler into thinking the variable is used.