Fix option already in use error message
It should check on p->opt not opt_table->opt

This old method was sometimes reporting the incorrect conflicting
option.

Also, the in use option list is printed which is also helpful. We could
take that out but it's what I used to fix the local source copy dsh vs
genders. We need both so we changed genders options in our internal pdsh
version.