slurm_plugin: Call slurm_init() once before any call to Slurm API
Since Slurm 20.11, slurm_init() needs to be called before any call
to the Slurm API to make sure the config file is read in.
We cannot call it in mod_slurm_init() or in mod_slurm_wcoll() directly
as it requires slurm to be configured before use, thus pdsh would
fail whenever the slurm plugin is found but Slurm is not configured
correctly.

Signed-off-by: Egbert Eich <eich@suse.com>