slurm: workaround Slurm 23.x new `list_t` typedef
Problem: Slurm exports a few generic container typedefs via slurm.h
like `List` and `ListIterator` which conflict with pdsh internal
definitions of the same types. To work around this, pdsh currently
defines `__list_datatypes_defined` in the slurm plugin to avoid Slurm
exporting those typedefs. However, in Slurm >23.x, a new typedef
is added, `list_t`, and the list handling functions are updated to
use this new type (which is actually just a new name for the same
type). Now pdsh can no longer use the `__list_datatypes_defined`
because then the necessary `list_t` data is not defined so including
the `slurm.h` header fails. However, without `__list_datatypes_defined`
compilation also fails because the `List` typedef is redefined.

Use the only available option and force the definitions of

 typedef struct xlist list_t;
 typedef struct listIterator list_itr_t;

in the pdsh slurm module before including `slurm.h`. This allows the
`__list_datatypes_defined` define to remain enabled, avoiding the
redefinition of `List` and `ListIterator`, while also ensuring that
the new names for these structs are defined.

This may be brittle in the long term, but is the only solution
available now.

Fixes #152